+++ images/lib/refinery/images/dragonfly.rb
@@ -8,6 +8,7 @@ module Refinery
         def setup!
           app_images = ::Dragonfly[:refinery_images]
           app_images.configure_with(:imagemagick)
+          app_images.configure_with(:rails)
 
           app_images.define_macro(::Refinery::Image, :image_accessor)
 
@@ -36,17 +37,12 @@ module Refinery
           end
         end
 
+        ##
+        # Injects Dragonfly::Middleware for Refinery::Images into the stack
         def attach!(app)
-          ### Extend active record ###
-          app.config.middleware.insert_before Refinery::Images.dragonfly_insert_before,
-                                              'Dragonfly::Middleware', :refinery_images
-
-          app.config.middleware.insert_before 'Dragonfly::Middleware', 'Rack::Cache', {
-            :verbose     => Refinery::Core.verbose_rack_cache,
-            :metastore   => "file:#{Rails.root.join('tmp', 'dragonfly', 'cache', 'meta')}",
-            :entitystore => "file:#{Rails.root.join('tmp', 'dragonfly', 'cache', 'body')}"
-          }
+          app.config.middleware.insert_after 'Rack::Cache', 'Dragonfly::Middleware', :refinery_images
         end
+
       end
 
     end