+++ changelog.md
@@ -15,6 +15,7 @@
 * Added Hungarian translations. [#2010](https://github.com/refinery/refinerycms/pull/2010). [minktom](https://github.com/minktom)
 * Extracted search header logic into partial. [#1974](https://github.com/refinery/refinerycms/pull/1974). [Uģis Ozols](https://github.com/ugisozols)
 * Images can only be updated when the image being uplaoded has the same filename as the original image. [#1866](https://github.com/refinery/refinerycms/pull/1866). [Philip Arndt](https://github.com/parndt) & [Uģis Ozols](https://github.com/ugisozols)
+* Fixed Rack::Cache middleware bug by making rack-cache a core dependency. Added Refinery::Cache module with extensive notes about caching configurations. [#1736](https://github.com/refinery/refinerycms/issues/1736). [Alexander Wenzowski](https://github.com/wenzowski)
 
 * [See full list](https://github.com/refinery/refinerycms/compare/2-0-stable...master)